WiX Installers
Migrated the installer to WiX. Since the current recommendation from Microsoft on their C++ Runtime is to install it from their exe, I split the installer up into a standalone

Note that the installed app doesn't run just yet. The MPIR dll that's in the repo is build against the v100 runtime, where HexEdit is build against, and this currently installs v142. Updating these dependencies is the next big step for me here.  The immediate next step is to remove the old installer files.

Also note that downloading the VCRuntime package is done as a build step in the standalone installer (the one that includes the runtime). I did this for convenience, but I'm not a fan of needing a network connection just to build, so I may revisit this later.

Other things done on this commit:
- Moved Templates up a folder. These aren't specific to the application code (maybe not even used by it directly), so un-burying these a bit.