Adding WixHelper
This is a small command line tool that will post-process sources generated from heat.  A better solution for this might be a HeatExtension (or a Mutator), but for the small bit I need to do here, this is the more-simple solution, at least to me right now.

This is a predecessor to integrating vcpkg as a package manager. I have this done on an experimental branch, but I'd like to clean up the history a bit so it makes a bit more sense if I ever have to look back at it.

This also removes crypto++ from the solution since I'm working off the completed change, and that project's getting deleted soon anyway. I cannot guarantee that this commit will build.