Removed global define of `_SECDLL`, added `_WIN32_IE` for debug builds.
`_SECDLL` seems to have been a holdover from ancient times (i.e: before the svn/git history). The only references to this macro I find online are related to an ObjectiveToolkit, that doesn't seem to be used even in the initial source import commit.

`_WIN32_IE` existed in the VS2008 project files (as well as the previous versions), so I assume its omission from the 2010 project file's debug configuration was a mistake.