Fixed a bad assert that triggered on redraw of MDI background.
`CMainFrame::OnEraseMDIClientBackground` had been asserting that `FreeImage_GetPixelColor` ran successfully. However, the input bitmap is a user-editable bitmap, so it could have failed without a coding error. Ultimately, the code handled failure as well, so having the assertion here seems pretty pointless. I'm guessing it was meant to detect if backgrnd.bmp was saved with the incorrect format, but since that's the bitmap it was triggering on, I'm not sure that it served its purpose if that was the case. The only history is that this assert was added on the v3.5 SVN branch, but any more details were lost in the git import.

I fixed this by generalizing the logic to get the pixel color. It tries to get the image's background color if it defines one, and if not, it reads the top-left pixel. The issue I ran into was that backgrnd.bmp uses indexed colors, so I added logic to check for (and handle) palette-based images as well.

As part of debugging, I added a FreeImage message output handler to TRACE out the message. It didn't end up helping here, but it doesn't seem like a bad thing to have around, so I am keeping it in.