First batch of testing (CFile64).
There's not much here, as I'm mostly just getting things set up and getting a feel for the catch2 test framework.  The tests added here are a basic set of tests for the instance functionality of `CFile64`, since this seemed like a pretty simple place to start.

As I was writing these few tests, I realized that CFile64 seems rather unnecessary given that MFC's regular CFile supports ULONGLONG. I dug into the history a bit, and found this class to be largely untouched since the SVN import in 2006. Digging a bit more online led me to some archived compuserve pages for WFC, which indeed talk about CFile64, and how it was actually made *for* HexEdit!

This did also confirm my suspicion that CFile64 was made in a time when MFC's CFile did *not* support ULONGLONG positions though.  However, I am a bit afraid to replace `CFile64` with the regular `CFile` at the moment - there's too high a chance of subtle differences that I would not catch.  There's also the question of whether `CFileNC` would still work if extending MFC's `CFile` as well.

Changes here:
- Added a new Tests project for testing. I settled on the catch2 framework for this after a little bit of searching. Since it's a header-only library (no runtime deps to mess up the main app), I threw it in the global vcpkg manifest.

- Added member variable initializer to CFile64's constructors. I noticed these didn't exist, and had some fun writing `Grabage.h` to verify that, yes, the members do contain whatever was leftover in memory without the initializers there. Digging through the archives the lack of member initializers stems from WFC's coding standards which say not to use them, possibly for an unclear C++ spec on when they are run. Anything unclear about that in the C++ spec has been cleared up.

- Removed CFile64::ReadHuge and WriteHuge, as these were not used, and just wrapped calls to the normal Read and Write methods.