Moved Graphics out to the resources directory.
This is a bit more normalization work before setting up a build pipeline. For that, I'm going to need the bmp/png/jpg files on the build environment. This means either that I'll need a way to convert the PDN files to these from a script (and the PDN devs seem unwilling to allow for this), or I'll need to commit the converted images. As much as I don't like it, the latter seems to be the best way forward. The (possibly even better) alternative would be to convert everything to a more script-friendly (and maybe even git-friendly) format, but I do not have the skills to do this.

On top of moving the files:
- Added a WiX variable for the graphics directory, and updated graphics file references in the installer sources to use it.
- Removed a reference to a seemingly unused bitmap from the HexEdit project that did not make it in to the original GitHub repo.
- Updated the gitignore for the new path and removed some obsolete entries.