Reconstructed banner.bmp's source PDN.
This also required me to extract the hextile source from About.pdn image, adjust the levels, apply a slight blur to make it more comprehensible, etc, etc... I threw the result into a new PDN file though so it's not lost. For the sake of being able to reproduce the banner source (just in case), I added a layer highlighting the exact portion of the image used.