Committing the image files.
This is in support of setting up an automated build pipeline. There isn't a great way to handle image resources in a git repo. That said, there's not a ton of image data here, and I don't anticipate changing these much, so they shouldn't bloat the repo _too_ much.

The ways to handle version controlling image files (and other binaries) all tend to involve either just living with the repo bloat, or moving images to a separate repo, usually a centralized one rather than distributed like git.  The separate repo seems like a bit much for how few image resources we have here, so I'm willing to just live with the bloat, at least for now.

The other factor here is that this now (or rather _still_) requires manually baking the PDN files into their needed image formats (the ones I'm committing here), but with PDN not supporting any sort of scripting, and needing at least some of these images to build the installer, I feel that committing the baked images is the only way to support an automated build pipeline. I don't particularly like it though...