Removed automatic update checks.
These checked against a hard-coded URL pointing at the old hexedit.com site. Since this site is defunct, the update check adds a significant (few seconds) delay on startup as it times out waiting for the connection.

I also added backgrnd.bmp and splash.bmp to the project such that they will be copied to the output directory on build. Without seeing the splash screen when debugging was making it rather difficult to see the delay the update check was causing (and how much removing the check reduced it).