Add missing "Open with HexEdit".
Moved this out of the separate "ContextMenu" feature; since the key references the installed exe's component, it must be part of the same feature.

I also compared this to a v4.0 installer and noticed that I also forget the check box to toggle the context menu option, so I added that here as well.

I also added a default value for the `RUNFROMBUNDLE` property since it was raising a warning (because the `<Property>` element wasn't actually doing anything before).