Tests for CReadSRecord. 1/?
First batch of tests for class `CReadSRecord`. While working on this, I found that the limitation on only being able to open real-deal files was going to make testing this a real pain. I left a TODO in the test file describing what I had hoped would be possible.

I wrote the tests on this commit and then made the refactor I describe in the comment (though with the stream wrapped in an `std::unique_ptr<CFile>` since MFC doesn't have move ctors), but I want to split that refactor into a separate commit just to avoid mixing the two kinds of changes.