Refactored CReadSRecord to accept a source stream from its ctor.
This is primarily to support testing, so for the moment, the original ctor is being kept around and preserved the existing behavior.

I updated the Get test case to test from both constructors and removed the TODO describing the refactor.

I also took this opportunity to change some inline member assignments in the ctor here to member initializers.