Refactor CReadSRecord into new HexImporter implementation.
The goal here is to possibly combine the srec/intel hex import functionality a bit. This is primarily as a means to learn the app code a bit more.

The desired outcome is to have a CHexEditView function that takes a `HexImporter&`, and have `do_intel` and `do_motorola` call that with their respective HexImporter implementations.  If this works out OK, then export will get the same treatment, and if they both work out, then I'll merge it all into main.

The only notable changes in functionality should be:
- the caller is now responsible for opening the stream instead of passing in a file path.
- changed win32 `BOOL` to standard `bool` type.

Changes here:
- Added HexImporter and SRecordImporter
- Refactored CReadSRecord into SRecordImporter
- Updated CHexEditView::do_motorola to use the new class.
- Updated CReadSRecordTests for the new class name
- Removed the "File not found" test since opening the stream is now done by the client.
- Added CHECKs for the new RecordsRead property.
- Removed test for null output buffer since it cannot happen under any normal circumstances. Since the buffer is allocated with operator new without std::nothrow (however that works with MFC's operator new macros), so if allocation fails it throws a CMemoryException rather than returns a null pointer.