Refactor CReadIntelHex into IntelHexImporter.
I did change one thing here:  When the import detects that a line is too short to be valid, the original code would skip that line. By mistake, I changed that to an error during the SREC import refactor. I noticed this here and fixed it, but realized that it meant that incomplete records could be skipped without notice.  The change is that I made this case an error if and only if the line begins with the correct initial character ('S' for SREC or ':' for Intel), and otherwise the line is skipped. I also added tests for either error/skip case.

Otherwise though, this is pretty much exactly the same deal as with the SREC refactor. I removed the same "construct from file path" and "buffer null" tests from the intel hex, I removed the original CReadIntelHex class, I updated CHexEditView::on_intel.