Additional tests for CWriteSRecord.
There's one change to the app code here, and that is an attempt to query if the stream is open in CWriteSRecord's destructor, which I added to avoid assertion failures from trying to write to a bad stream.  When I create the corresponding HexExporter class, I think I'm going to change this to have separate `WritePrologue` and `WriteEpilogue` functions so I don't need to query this from the dtor just to avoid writing the S5 to a bad stream...

I also added some more failure points to CErrorFile and moved it out to a shared file.