[WIP] cryptography refactor 1/?
This adds the first pass at the new design I'm looking for. I'm
hoping this Provider/Algorithm design will help make this a bit more
portable/extensible in the future.

I am doing this refactor a bit backwards though - I am not in the
slightest bit familiar with the Windows Crypt APIs, aside from seeing
the docs say they're deprecated in favor of the BCrypt ones. Doing
this design first is intended for learning these APIs, and now that
this is in place, I'm planning to write tests for the existing code to
ensure I'm not breaking anything user-visible.

In the future, I'm hoping to add providers for the new BCrypt APIs in
Windows, and probably also the Crypto++ library we're already
referencing here.

Even further in the future though, I'm hoping to port this to other
platforms (probably using wxWidgets for UI), and this design should
let me more easily swap in/out providers based on the target platform.
This is also why I put the ADVAPI provider in a `windows` directory
here.