Tests for existing CCrypto class.
This makes a couple changes to the code to enable testing:
- added a ctor to take a dialog provider to test branches that show
  message boxes

- Removed some `ASSERT(0)`s on branches that can, at least in theory,
  be reached if ADVAPI32.dll either doesn't exist or has some
  improperly configured algorithms; as well as some on branches that
  are otherwise handled properly.

With these changes, the only uncovered lines are for
`LoadLibrary`/`GetProcAddress` failing, and being unable to create the
hash algorithm or key objects. ADVAPI32.dll is part of Windows, so I'm
not sure it's possible to have the load fail, and the functions we're
loading here are all documented, so the `GetProcAddress` calls should
always succeed as well. I'm not sure if there's any way to create
custom CSPs, so I don't know if it's possible to inject errors in the
other cases here, but this code should check for errors here
regardless.

I also include a couple tests comparing the ADVAPI32 implementation to
that provided by Crypto++. For the sake of normalizing the available
encryption algorithms, as well as not relying on Windows APIs as much,
I'm hoping to add (and maybe later, migrate to) the algorithm impls in
Crypto++, and these tests are intended to roughly verify they work the
same.