A little bit of cleanup in OnEncrypt/OnDecrypt.
I managed to move the identical password prompts in both up to a new
function on CHexEditApp with the rest of the cryptography funcs there,
but these two functions seem to be _just_ different enough that it'd
be difficult to combine any more.

Instead, I did a pass over the two functions to try and normalize
things between them a bit - things like messages and when the wait
cursor is shown. I also changed the raw pointer for the buffer to a
unique_ptr in both places while I was in there.

The change to TParser just fixes a warning on passing a CString for
%s by adding a cast.