Add Tests for Misc time functions.
I also removed some functions that were either no-longer used
(`ConvertToFileTime`) or were #ifdef'd out (`SetFileCreationTime`
and `SetFileAccessTime`). The latter functions were used for the
registration, and I have no intent to bring that back and monetize
someone else's work here, so away they go.

Since one of these tests touched the C locale, I also wanted a RAII
type for restoring that so it's not forgotten on (eg) an exception, so
I added the GSL as a dependency for `gsl::finally`/`gsl::final_action`.