Refactor of strtoi64.
The three-arg overload matches the now-standard std::strtoll for all
relevant inputs that I could think of, so I just replaced calls to
that overload to strtoll. The only two differences are:
1. How errors are reported: the original impl only returned endptr as
   something other than a ptr to the NUL terminator (since the only
   strings parsed here are *just* the numbers), and the standard one
   returns errors through errno.

2. Handling of the top-most (sign) bit: strtoll will return failure
   when the magnitude of the value is too large, but no sign exists.
   The previous impl seems to have happily read the top-most bit as
   part of the value though. I believe this effects 64-bit values
   where the top-most digit is 8-F. In places where it matters, I
   instead read as a u64, and just cast the bits to s64.

The two-arg overload matches strtoll, except that it skips any invalid
characters instead of failing immediately. Changed to allocate a small
buffer locally, copy only the valid digits to there, and then parse
that using the standard function. This is ultimately sort of a
side-step, but it's one less number parsing function I have to care
about here.