Tests for FindFirstDiff.
This includes a couple benchmarks for it as well, since I was curious
just how much of an improvement the XMM scan have over a simple byte
scan that's been through the optimizer, and it turns out its still
about 10x faster!

When I ran with optimizations enabled though, I noticed some other
tests are failing outside of debug builds, so need to get those fixed.
And add a test step to the CI pipeline, which I thought I already had
in there...