Update pipeline for benchmarks.
The one benchmarking case takes about 8x as long as the actual tests;
around 4200ms compared to 500ms for me locally. It's obviously not a
huge amount of time right now, but there are some things I know I'm
going to want more benchmarks for (esp. in template processing), so I
want to split this off as a separate, manually-enabled step on the
pipeline before the time grows too much.

Of course, that also means I'm probably going to want to deal with
the build times at some point, but this is just the low-hanging fruit
right now. Plus, now that the vcpkg baseline seems to be working, the
build time's down to around 12 minutes (from nearly an hour) on the
pipeline.